new file:   fWapp.py; renamed:    MainWindow.py -> fWlayout.py; renamed:    mainwindow.ui -> fWlayout.ui; modified:   freeWrite.py; new file:   icon.png
