+++ freeWrite.py
@@ -1,68 +1,19 @@
 #!/usr/bin/env python3.7
 # -*- coding: utf-8 -*-
 """
-Created on Sun Jul 21 12:35:30 2019
 
-@author: raydialow
+freeWrite 
+
+An application for the collection and archival of personal logs. The application should encourage daily logging and review. The application should discourage deletion or revision.
+
+author contact: raydialow@raydialow.xyz
+
 """
-
-import sys, dbm
+import sys
 from PyQt5 import QtWidgets
-from PyQt5.QtCore import QDate
-from MainWindow import Ui_MainWindow # MainWindow.py generated by QtDesigner
+from fWapp import MainWindow
 
-# CONFIG TODO: Import from config file...
-dbfilename = "freeWrite.db"
-dailygoal = 500 # presently in bytes TODO in words
-debug = True
-
-# MAIN WINDOW 
-class MainWindow(QtWidgets.QMainWindow, Ui_MainWindow):
-    # CONSTRUCTOR
-    def __init__(self, *args, obj=None, **kwargs):
-        super(MainWindow, self).__init__(*args, **kwargs)
-        self.setupUi(self)
-        self.readDate()
-        self.Today = QDate.currentDate().toString()
-        print("Today is " + self.Today)
-        self.progressBar.setMinimum(0)
-        self.progressBar.setMaximum(dailygoal)
-        #SIGNALS TODO
-        self.calendarWidget.selectionChanged.connect(self.readDate)
-        self.lineEdit.returnPressed.connect(self.appendEntry)
-    
-    # SLOTS TODO
-    def readDate(self): # Uses selected date to update text browser
-        selDate = self.calendarWidget.selectedDate().toString()
-        if(debug): print("Read: " + selDate)
-        selText = ""
-        with dbm.open(dbfilename,'c') as db:
-            try: selText = db[selDate]
-            except: selText = b""
-            finally: 
-                self.textBrowser.setPlainText(str(selText,'utf-8'))
-                self.progressBar.setValue(len(selText))
-                
-    def appendEntry(self):
-        selDate = self.calendarWidget.selectedDate().toString()
-        if(selDate in self.Today):
-            if(debug): print("Write: " + selDate)
-            addText = self.lineEdit.text()
-            with dbm.open(dbfilename,'c') as db:
-                try: db[selDate] = db[selDate] + bytes(addText,'utf-8') + b"\n"
-                except: db[selDate] = self.lineEdit.text() + b"\n"
-                finally: self.lineEdit.clear()
-            self.readDate()
-        elif(debug): 
-            print("Prohibited: " + selDate + " is not Today " + self.Today)
-        else: pass
-    
-
-# INSTANCE
 app = QtWidgets.QApplication(sys.argv)
 window = MainWindow()
 window.show()
-
-# EXECUTE
 app.exec()
-
